app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
websockets>=12.0
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
aiomysql>=0.2.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
firebase_admin
onnxruntime>=1.16.0
numpy>=1.26.0